+++ CHANGELOG.md
@@ -12,6 +12,9 @@ Latest
  * Removed sqlalchemy as a requirement for SunPy
  * Added support for NOAA solar cycle prediction in lightcurves
  * Fix algorithm in sunpy.sun.equation_of_center
+ * Added Docstrings to LightCurve methods.
+ * Added functions that implement image coalignment with support for MapCubes. 
+ * Added contains functionality to TimeRange module
 
 0.4.0
 -----